fix: chart preview
